deleted unused, more refactor in WatchlistsPopup
