--- src/components/WatchlistPopup/Watchlists.css
@@ -1,81 +1,0 @@
-.watchlists {
-  position: relative;
-  display: flex;
-  flex-direction: column;
-  min-width: 350px;
-}
-
-.watchlists .watchlists__empty-list-msg {
-  margin: 0 auto;
-  margin-top: 20px;
-  text-align: center;
-  font-weight: bold;
-}
-
-.watchlists__header {
-  display: flex;
-  justify-content: flex-end;
-  padding: 5px 0;
-}
-
-.watchlists .watchlists__list {
-  margin-bottom: 10px;
-  min-height: 100px;
-}
-
-.watchlists__list .watchlists__item {
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  min-height: 2em;
-  border-bottom: 1px solid #eee;
-}
-
-.watchlists__list .watchlists__item__link {
-  display: flex;
-  flex: 1;
-  align-items: center;
-  justify-content: space-between;
-  padding: 8px 15px;
-  cursor: pointer;
-  min-height: 2em;
-  border-bottom: none;
-}
-
-.watchlists__list .watchlists__item__link:hover {
-  background: #d4d9dd5e;
-}
-
-.watchlists__list .watchlists__item__link:first-child {
-  border-top: none;
-}
-
-.watchlists__tools .icon {
-  cursor: pointer;
-  margin: 5px;
-  margin-left: 0;
-  margin-right: 10px;
-}
-
-.watchlists__tools .icon-green {
-  color: green;
-}
-
-.watchlists__item__name {
-  display: flex;
-}
-
-.watchlists__item__name div {
-  display: flex;
-  padding-top: 4px;
-  text-transform: uppercase;
-}
-
-.watchlists__tools__move-to-trash {
-  opacity: 0.1 !important;
-  transition: all 0.2s ease;
-}
-
-.watchlists__tools__move-to-trash:hover {
-  opacity: 1 !important;
-}
